services/backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
aerich >= 0.7.0
asyncpg
passlib[bcrypt]==1.7.4
tortoise-orm==0.19.2
uvicorn==0.20.0
pytest==8.2.2
pytest-asyncio==0.23.8
python-jose==3.3.0
python-multipart==0.0.5